Implementation/Documentation
Added option whether or not to create the plugin in an organization
(creates under your user instead if you do not select). Also gave option
for whether or not to integrate with HipChat. Provided documentation on
how to use the plugin in conjunction with the bundled XL Release
template in both the README as well as in the template itself as
markdown in the new release page and in manual tasks within the release
itself.