Template/Required Fields Fix
Updated the template to include a manual step for cloning the repository
to the local machine after creation to complete the codacy step.

Defaulted the value of the "Create In GitHub Organization" value to the
CREATE_IN_ORG value defined on release creation.

Updated GitHub Organization, HipChat Token, and HipChat Room ID to all
be non-required values. You only need to define them if you choose to
create the repository in an organization, and if you decide you want to
integrate with HipChat.